updated domain classes, introduced strategy pattern, updated EJB and EJBImpl classes, updated controllers

+++ src/main/java/ejb/impl/ArrivalEJBImpl.java
@@ -7,6 +7,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Stateless
@@ -21,21 +22,35 @@ public class ArrivalEJBImpl implements ArrivalEJB {
     }
 
     @Override
-    public void create(Arrival arrival) {
+    public Arrival findById(Long id) {
+        return em.find(Arrival.class, id);
+    }
+
+    @Override
+    public Response create(Arrival arrival) {
         em.persist(arrival);
+        return Response.ok("Arrival created").build();
     }
 
     @Override
-    public void update(Long id, Arrival body) {
+    public Response update(Long id, Arrival body) {
         Arrival arrival = em.find(Arrival.class, id);
-        arrival.setPort(body.getPort());
-        arrival.setDate(body.getDate());
-        em.merge(arrival);
+        if (arrival != null) {
+            arrival.setPort(body.getPort());
+            arrival.setDate(body.getDate());
+            em.merge(arrival);
+            return Response.ok("Arrival updated").build();
+        }
+        return Response.status(404).build();
     }
 
     @Override
-    public void remove(Long id) {
+    public Response remove(Long id) {
         Arrival arrival = em.find(Arrival.class, id);
-        em.remove(arrival);
+        if (arrival != null) {
+            em.remove(arrival);
+            return Response.ok("Arrival removed").build();
+        }
+        return Response.status(404).build();
     }
 }
